requirements.txt
monai-weekly[nibabel, tqdm]
matplotlib
click
pytest
flake8
invoke